password updates and account delete with securely

+++ src/app/api/users/[userId]/route.js
@@ -18,7 +18,19 @@ export async function GET(request, { params }) {
 
 export async function DELETE(request, { params }) {
   const { userId } = params;
+  const { password } = await request.json();
+
   try {
+    const user = await User.findById(userId);
+    if (!user) {
+      return NextResponse.json({ message: "User not found", success: false }, { status: 404 });
+    }
+
+    const isMatch = await bcrypt.compare(password, user.password);
+    if (!isMatch) {
+      return NextResponse.json({ message: "Password is incorrect", success: false }, { status: 400 });
+    }
+
     await User.deleteOne({ _id: userId });
     return NextResponse.json({ message: "User deleted successfully", success: true });
   } catch (error) {
@@ -29,7 +41,7 @@ export async function DELETE(request, { params }) {
 
 export async function PUT(request, { params }) {
   const { userId } = params;
-  const { username, password, profileURL } = await request.json();
+  const { username, oldPassword, newPassword, profileURL } = await request.json();
 
   try {
     const user = await User.findById(userId);
@@ -43,9 +55,12 @@ export async function PUT(request, { params }) {
     if (profileURL) {
       user.profileURL = profileURL;
     }
-    if (password) {
-      // Hash the new password with a sufficient number of salt rounds
-      const hashedPassword = await bcrypt.hash(password, 10);
+    if (newPassword) {
+      const isMatch = await bcrypt.compare(oldPassword, user.password);
+      if (!isMatch) {
+        return NextResponse.json({ message: "Old password is incorrect", success: false }, { status: 400 });
+      }
+      const hashedPassword = await bcrypt.hash(newPassword, 10);
       user.password = hashedPassword;
     }
